path_finder/requirements.txt
streamlit==1.22.0
networkx==3.1
matplotlib==3.7.1
folium==0.14.0
pandas==2.0.1
numpy==1.24.3
googlemaps==4.10.0
python-dotenv==1.0.0
fpdf==1.7.2
scikit-learn==1.2.2
deap==1.3.3
requests==2.30.0
geopy==2.3.0 